requirements.txt
# Core Static Analysis
slither-analyzer>=0.10.0    # Primary static analysis framework for Solidity
solc-select>=1.0.4          # Manage multiple Solidity compiler versions

# Machine Learning
scikit-learn>=1.3.0         # Classical ML: Random Forest, Logistic Regression
numpy>=1.24.0               # Numerical operations for feature vectors
pandas>=2.0.0               # Dataset manipulation and CSV handling

# Graph Analysis
networkx>=3.1               # Call graph construction and cycle detection

# CLI Interface
click>=8.1.0                # Modern CLI framework with decorators
rich>=13.0.0                # Beautiful terminal output (tables, colors, progress bars)
colorama>=0.4.6             # Cross-platform colored terminal text

# Data Processing
pyyaml>=6.0                 # Configuration file parsing
jsonschema>=4.17.0          # Validate JSON outputs

# Development & Testing
pytest>=7.4.0               # Unit testing framework
pytest-cov>=4.1.0           # Code coverage reports
black>=23.0.0               # Code formatter
flake8>=6.0.0               # Linting

# Utilities
tqdm>=4.65.0                # Progress bars for batch processing
python-dotenv>=1.0.0        # Environment variable management
